+++ python/ray/tests/BUILD
@@ -95,7 +95,6 @@ py_test_module_list(
     "test_threaded_actor.py",
     "test_tls_auth.py",
     "test_ray_debugger.py",
-    "test_worker_capping.py",
   ],
   size = "medium",
   extra_srcs = SRCS,
@@ -214,7 +213,6 @@ py_test_module_list(
     "test_object_manager.py",
     "test_multi_tenancy.py",
     "test_scheduling.py",
-    "test_scheduling_2.py",
   ],
   size = "medium",
   extra_srcs = SRCS,
@@ -392,7 +390,6 @@ py_test_module_list(
     "test_object_spilling_2.py",
     "test_object_spilling_3.py",
     "test_scheduling.py",
-    "test_scheduling_2.py",
     "test_array.py"
   ],
   size = "medium",
@@ -408,7 +405,6 @@ py_test_module_list(
     "test_object_spilling_2.py",
     "test_object_spilling_3.py",
     "test_scheduling.py",
-    "test_scheduling_2.py",
     "test_array.py"
   ],
   size = "large",
