--- python/ray/tests/test_scheduling_2.py
@@ -1,58 +1,0 @@
-import numpy as np
-import platform
-import pytest
-import sys
-import time
-
-import ray
-
-
-@pytest.mark.skipif(
-    platform.system() == "Windows", reason="Failing on Windows. Multi node.")
-def test_load_balancing_under_constrained_memory(ray_start_cluster):
-    # This test ensures that tasks are being assigned to all raylets in a
-    # roughly equal manner even when the tasks have dependencies.
-    cluster = ray_start_cluster
-    num_nodes = 3
-    num_cpus = 4
-    object_size = 4e7
-    num_tasks = 100
-    for _ in range(num_nodes):
-        cluster.add_node(
-            num_cpus=num_cpus,
-            memory=(num_cpus - 2) * object_size,
-            object_store_memory=(num_cpus - 2) * object_size)
-    cluster.add_node(
-        num_cpus=0,
-        resources={"custom": 1},
-        memory=(num_tasks + 1) * object_size,
-        object_store_memory=(num_tasks + 1) * object_size)
-    ray.init(address=cluster.address)
-
-    @ray.remote(num_cpus=0, resources={"custom": 1})
-    def create_object():
-        return np.zeros(int(object_size), dtype=np.uint8)
-
-    @ray.remote
-    def f(i, x):
-        print(i, ray.worker.global_worker.node.unique_id)
-        time.sleep(0.1)
-        return ray.worker.global_worker.node.unique_id
-
-    deps = [create_object.remote() for _ in range(num_tasks)]
-    for i, dep in enumerate(deps):
-        print(i, dep)
-
-    # TODO(swang): Actually test load balancing. Load balancing is currently
-    # flaky on Travis, probably due to the scheduling policy ping-ponging
-    # waiting tasks.
-    deps = [create_object.remote() for _ in range(num_tasks)]
-    tasks = [f.remote(i, dep) for i, dep in enumerate(deps)]
-    for i, dep in enumerate(deps):
-        print(i, dep)
-    ray.get(tasks)
-
-
-if __name__ == "__main__":
-    import pytest
-    sys.exit(pytest.main(["-v", __file__]))
