+++ src/ray/common/ray_config_def.h
@@ -106,23 +106,17 @@ RAY_CONFIG(int64_t, max_lineage_bytes, 1024 * 1024 * 1024)
 /// See also: https://github.com/ray-project/ray/issues/14182
 RAY_CONFIG(bool, preallocate_plasma_memory, false)
 
-/// We place a soft cap on the number of tasks of a given scheduling class that
-/// can run at once to limit the total nubmer of worker processes. After the
-/// specified interval, the new task above that cap is allowed to run. The time
-/// before the next tasks (above the cap) are allowed to run increases
-/// exponentially. The soft cap is needed to prevent deadlock in the case where
-/// a task begins to execute and tries to `ray.get` another task of the same
-/// class.
-RAY_CONFIG(int64_t, worker_cap_initial_backoff_delay_ms, 1000)
-
-/// After reaching the worker cap, the backoff delay will grow exponentially,
-/// until it hits a maximum delay.
-RAY_CONFIG(int64_t, worker_cap_max_backoff_delay_ms, 1000 * 10)
-
 /// The fraction of resource utilization on a node after which the scheduler starts
 /// to prefer spreading tasks to other nodes. This balances between locality and
 /// even balancing of load. Low values (min 0.0) encourage more load spreading.
 RAY_CONFIG(float, scheduler_spread_threshold, 0.5);
+
+// TODO (Alex): Remove this feature flag once scheduling class capping is
+// implemented.
+/// Whether to include function descriptors, and depth in the
+// scheduling class. / This causes tasks to be queued differently, so it may
+// effect scheduling / behavior.
+RAY_CONFIG(bool, complex_scheduling_class, false)
 
 // The max allowed size in bytes of a return object from direct actor calls.
 // Objects larger than this size will be spilled/promoted to plasma.
