+++ src/ray/common/task/scheduling_resources.cc
@@ -189,11 +189,6 @@ const ResourceSet ResourceSet::GetNumCpus() const {
   return cpu_resource_set;
 }
 
-double ResourceSet::GetNumCpusAsDouble() const {
-  const FixedPoint cpu_quantity = GetResource(kCPU_ResourceLabel);
-  return cpu_quantity.Double();
-}
-
 std::string format_resource(std::string resource_name, double quantity) {
   if (resource_name == "object_store_memory" ||
       resource_name.find(kMemory_ResourceLabel) == 0) {
