+++ src/ray/raylet/node_manager.cc
@@ -355,10 +355,6 @@ NodeManager::NodeManager(instrumented_io_context &io_service, const NodeID &self
             self_node_id_, resource_names, nullptr));
       });
 
-  periodical_runner_.RunFnPeriodically(
-      [this]() { cluster_task_manager_->ScheduleAndDispatchTasks(); },
-      RayConfig::instance().worker_cap_initial_backoff_delay_ms());
-
   RAY_CHECK_OK(store_client_.Connect(config.store_socket_name.c_str()));
   // Run the node manger rpc server.
   node_manager_server_.RegisterService(node_manager_service_);
