+++ src/ray/raylet/scheduling/cluster_resource_scheduler.cc
@@ -561,13 +561,6 @@ std::string ClusterResourceScheduler::DebugString(void) const {
   return buffer.str();
 }
 
-uint64_t ClusterResourceScheduler::GetNumCpus() const {
-  auto it = nodes_.find(local_node_id_);
-  RAY_CHECK(it != nodes_.end());
-  return static_cast<uint64_t>(
-      it->second.GetLocalView().predefined_resources[CPU].total.Double());
-}
-
 void ClusterResourceScheduler::InitResourceInstances(
     FixedPoint total, bool unit_instances, ResourceInstanceCapacities *instance_list) {
   if (unit_instances) {
