+++ src/ray/raylet/scheduling/cluster_task_manager.h
@@ -136,18 +136,7 @@ class ClusterTaskManager : public ClusterTaskManagerInterface {
   /// \param task_dependency_manager_ Used to fetch task's dependencies.
   /// \param is_owner_alive: A callback which returns if the owner process is alive
   /// (according to our ownership model).
-  /// \param get_node_info: Function that returns the node info for a node.
-  /// \param announce_infeasible_task: Callback that invokes the user if a task
-  ///        is infeasible.
-  /// \param worker_pool: A reference to the worker pool.
-  /// \param leased_workers: A reference to the leased workers map.
-  /// \param get_task_arguments: A callback for getting a tasks' arguments by
-  ///        their ids.
-  /// \param max_pinned_task_arguments_bytes: The cap on pinned arguments.
-  /// \param get_time_ms: A callback which returns the current time in milliseconds.
-  /// \param sched_cls_cap_interval_ms: The time before we increase the cap
-  ///        on the number of tasks that can run per scheduling class. If set
-  ///        to 0, there is no cap. If it's a large number, the cap is hard.
+  /// \param gcs_client: A gcs client.
   ClusterTaskManager(
       const NodeID &self_node_id,
       std::shared_ptr<ClusterResourceScheduler> cluster_resource_scheduler,
@@ -160,11 +149,7 @@ class ClusterTaskManager : public ClusterTaskManagerInterface {
       std::function<bool(const std::vector<ObjectID> &object_ids,
                          std::vector<std::unique_ptr<RayObject>> *results)>
           get_task_arguments,
-      size_t max_pinned_task_arguments_bytes,
-      std::function<int64_t(void)> get_time_ms =
-          []() { return (int64_t)(absl::GetCurrentTimeNanos() / 1e6); },
-      int64_t sched_cls_cap_interval_ms =
-          RayConfig::instance().worker_cap_initial_backoff_delay_ms());
+      size_t max_pinned_task_arguments_bytes);
 
   void SetWorkerBacklog(SchedulingClass scheduling_class, const WorkerID &worker_id,
                         int64_t backlog_size) override;
@@ -272,8 +257,6 @@ class ClusterTaskManager : public ClusterTaskManagerInterface {
   ResourceSet CalcNormalTaskResources() const override;
 
  private:
-  struct SchedulingClassInfo;
-
   /// (Step 2) For each task in tasks_to_schedule_, pick a node in the system
   /// (local or remote) that has enough resources available to run the task, if
   /// any such node exist. Skip tasks which are not schedulable.
@@ -315,14 +298,6 @@ class ClusterTaskManager : public ClusterTaskManagerInterface {
   // queue.
   void SpillWaitingTasks();
 
-  /// Calculate the maximum number of running tasks for a given scheduling
-  /// class. https://github.com/ray-project/ray/issues/16973
-  ///
-  /// \param sched_cls_id The scheduling class in question.
-  /// \returns The maximum number instances of that scheduling class that
-  ///          should be running (or blocked) at once.
-  uint64_t MaxRunningTasksPerSchedulingClass(SchedulingClass sched_cls_id) const;
-
   /// Helper method to get the best node for running the task.
   std::string GetBestSchedulableNode(const internal::Work &work,
                                      bool requires_object_store_memory,
@@ -348,27 +323,6 @@ class ClusterTaskManager : public ClusterTaskManagerInterface {
   /// Tasks move from scheduled -> dispatch | waiting.
   absl::flat_hash_map<SchedulingClass, std::deque<std::shared_ptr<internal::Work>>>
       tasks_to_schedule_;
-
-  /// Tracking information about the currently running tasks in a scheduling
-  /// class. This information is used to place a cap on the number of running
-  /// running tasks per scheduling class.
-  struct SchedulingClassInfo {
-    SchedulingClassInfo(int64_t cap)
-        : running_tasks(),
-          capacity(cap),
-          next_update_time(std::numeric_limits<int64_t>::max()) {}
-    /// Track the running task ids in this scheduling class.
-    absl::flat_hash_set<TaskID> running_tasks;
-    /// The total number of tasks that can run from this scheduling class.
-    const uint64_t capacity;
-    /// The next time that a new task of this scheduling class may be dispatched.
-    int64_t next_update_time;
-  };
-
-  /// Mapping from scheduling class to information about the running tasks of
-  /// the scheduling class. See `struct SchedulingClassInfo` above for more
-  /// details about what information is tracked.
-  absl::flat_hash_map<SchedulingClass, SchedulingClassInfo> info_by_sched_cls_;
 
   /// Queue of lease requests that should be scheduled onto workers.
   /// Tasks move from scheduled | waiting -> dispatch.
@@ -440,14 +394,6 @@ class ClusterTaskManager : public ClusterTaskManagerInterface {
 
   /// The maximum amount of bytes that can be used by executing task arguments.
   size_t max_pinned_task_arguments_bytes_;
-
-  /// Returns the current time in milliseconds.
-  std::function<int64_t()> get_time_ms_;
-
-  /// The initial interval before the cap on the number of worker processes is increased.
-  const int64_t sched_cls_cap_interval_ms_;
-
-  const int64_t sched_cls_cap_max_ms_;
 
   /// Metrics collected since the last report.
   uint64_t metric_tasks_queued_;
