+++ src/ray/raylet/scheduling/cluster_task_manager_test.cc
@@ -225,8 +225,7 @@ class ClusterTaskManagerTest : public ::testing::Test {
               }
               return true;
             },
-            /*max_pinned_task_arguments_bytes=*/1000,
-            /*get_time=*/[this]() { return current_time_ms_; }) {}
+            /*max_pinned_task_arguments_bytes=*/1000) {}
 
   void SetUp() {
     static rpc::GcsNodeInfo node_info;
@@ -263,9 +262,7 @@ class ClusterTaskManagerTest : public ::testing::Test {
     ASSERT_TRUE(task_manager_.infeasible_tasks_.empty());
     ASSERT_TRUE(task_manager_.executing_task_args_.empty());
     ASSERT_TRUE(task_manager_.pinned_task_arguments_.empty());
-    ASSERT_TRUE(task_manager_.info_by_sched_cls_.empty());
     ASSERT_EQ(task_manager_.pinned_task_arguments_bytes_, 0);
-    ASSERT_TRUE(task_manager_.info_by_sched_cls_.empty());
     ASSERT_TRUE(dependency_manager_.subscribed_tasks.empty());
   }
 
@@ -303,7 +300,6 @@ class ClusterTaskManagerTest : public ::testing::Test {
   int node_info_calls_;
   int announce_infeasible_task_calls_;
   absl::flat_hash_map<NodeID, rpc::GcsNodeInfo> node_info_;
-  int64_t current_time_ms_ = 0;
 
   MockTaskDependencyManager dependency_manager_;
   ClusterTaskManager task_manager_;
@@ -358,59 +354,6 @@ TEST_F(ClusterTaskManagerTest, BasicTest) {
   task_manager_.TaskFinished(leased_workers_.begin()->second, &finished_task);
   ASSERT_EQ(finished_task.GetTaskSpecification().TaskId(),
             task.GetTaskSpecification().TaskId());
-  AssertNoLeaks();
-}
-
-TEST_F(ClusterTaskManagerTest, IdempotencyTest) {
-  /*
-    A few task manager methods are meant to be idempotent.
-    * `TaskFinished`
-    * `ReleaseCpuResourcesFromUnblockedWorker`
-    * `ReturnCpuResourcesToBlockedWorker`
-   */
-  RayTask task = CreateTask({{ray::kCPU_ResourceLabel, 4}});
-  rpc::RequestWorkerLeaseReply reply;
-  bool callback_occurred = false;
-  bool *callback_occurred_ptr = &callback_occurred;
-  auto callback = [callback_occurred_ptr](Status, std::function<void()>,
-                                          std::function<void()>) {
-    *callback_occurred_ptr = true;
-  };
-
-  task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  pool_.TriggerCallbacks();
-  ASSERT_FALSE(callback_occurred);
-  ASSERT_EQ(leased_workers_.size(), 0);
-  ASSERT_EQ(pool_.workers.size(), 0);
-
-  std::shared_ptr<MockWorker> worker =
-      std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234);
-  pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker));
-  pool_.TriggerCallbacks();
-
-  ASSERT_TRUE(callback_occurred);
-  ASSERT_EQ(leased_workers_.size(), 1);
-  ASSERT_EQ(pool_.workers.size(), 0);
-  ASSERT_EQ(node_info_calls_, 0);
-
-  ASSERT_EQ(scheduler_->GetLocalAvailableCpus(), 4.0);
-
-  task_manager_.ReleaseCpuResourcesFromUnblockedWorker(worker);
-  task_manager_.ReleaseCpuResourcesFromUnblockedWorker(worker);
-
-  ASSERT_EQ(scheduler_->GetLocalAvailableCpus(), 8.0);
-
-  task_manager_.ReturnCpuResourcesToBlockedWorker(worker);
-  task_manager_.ReturnCpuResourcesToBlockedWorker(worker);
-
-  ASSERT_EQ(scheduler_->GetLocalAvailableCpus(), 4.0);
-
-  RayTask finished_task;
-  task_manager_.TaskFinished(leased_workers_.begin()->second, &finished_task);
-  task_manager_.TaskFinished(leased_workers_.begin()->second, &finished_task);
-  ASSERT_EQ(finished_task.GetTaskSpecification().TaskId(),
-            task.GetTaskSpecification().TaskId());
-  ASSERT_EQ(scheduler_->GetLocalAvailableCpus(), 8.0);
   AssertNoLeaks();
 }
 
@@ -1133,7 +1076,6 @@ TEST_F(ClusterTaskManagerTest, OwnerDeadTest) {
   ASSERT_FALSE(callback_occurred);
   ASSERT_EQ(leased_workers_.size(), 0);
   ASSERT_EQ(pool_.workers.size(), 1);
-
   AssertNoLeaks();
 }
 
@@ -1682,400 +1624,6 @@ TEST_F(ClusterTaskManagerTest, PopWorkerExactlyOnce) {
   AssertNoLeaks();
 }
 
-TEST_F(ClusterTaskManagerTest, CapRunningOnDispatchQueue) {
-  scheduler_->AddLocalResourceInstances(ray::kGPU_ResourceLabel, {1, 1, 1});
-  RayTask task = CreateTask({{ray::kCPU_ResourceLabel, 4}, {ray::kGPU_ResourceLabel, 1}},
-                            /*num_args=*/0, /*args=*/{});
-  RayTask task2 = CreateTask({{ray::kCPU_ResourceLabel, 4}, {ray::kGPU_ResourceLabel, 1}},
-                             /*num_args=*/0, /*args=*/{});
-  RayTask task3 = CreateTask({{ray::kCPU_ResourceLabel, 4}, {ray::kGPU_ResourceLabel, 1}},
-                             /*num_args=*/0, /*args=*/{});
-  auto runtime_env_hash = task.GetTaskSpecification().GetRuntimeEnvHash();
-  std::vector<std::shared_ptr<MockWorker>> workers;
-  for (int i = 0; i < 3; i++) {
-    std::shared_ptr<MockWorker> worker =
-        std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-    pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker));
-    pool_.TriggerCallbacks();
-    workers.push_back(worker);
-  }
-  rpc::RequestWorkerLeaseReply reply;
-  int num_callbacks = 0;
-  auto callback = [&num_callbacks](Status, std::function<void()>, std::function<void()>) {
-    num_callbacks++;
-  };
-  task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  task_manager_.QueueAndScheduleTask(task2, false, &reply, callback);
-  task_manager_.QueueAndScheduleTask(task3, false, &reply, callback);
-  pool_.TriggerCallbacks();
-
-  ASSERT_EQ(num_callbacks, 2);
-
-  task_manager_.ReleaseCpuResourcesFromUnblockedWorker(workers[0]);
-  task_manager_.ReleaseCpuResourcesFromUnblockedWorker(workers[1]);
-
-  task_manager_.ScheduleAndDispatchTasks();
-
-  // Even though there are free resources, we've hit our cap of (8/4=)2 workers
-  // of the given scheduling class so we shouldn't dispatch the remaining task.
-  ASSERT_EQ(num_callbacks, 2);
-
-  RayTask buf;
-  task_manager_.TaskFinished(workers[1], &buf);
-
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-  ASSERT_EQ(num_callbacks, 3);
-
-  task_manager_.TaskFinished(workers[0], &buf);
-  task_manager_.TaskFinished(workers[2], &buf);
-
-  AssertNoLeaks();
-}
-
-TEST_F(ClusterTaskManagerTest, ZeroCPUTasks) {
-  scheduler_->AddLocalResourceInstances(ray::kGPU_ResourceLabel, {1, 1, 1});
-  RayTask task = CreateTask({{"GPU", 1}}, /*num_args=*/0, /*args=*/{});
-  RayTask task2 = CreateTask({{"GPU", 1}}, /*num_args=*/0, /*args=*/{});
-  RayTask task3 = CreateTask({{"GPU", 1}}, /*num_args=*/0, /*args=*/{});
-  auto runtime_env_hash = task.GetTaskSpecification().GetRuntimeEnvHash();
-  std::vector<std::shared_ptr<MockWorker>> workers;
-  for (int i = 0; i < 3; i++) {
-    std::shared_ptr<MockWorker> worker =
-        std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-    pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker));
-    pool_.TriggerCallbacks();
-    workers.push_back(worker);
-  }
-  rpc::RequestWorkerLeaseReply reply;
-  int num_callbacks = 0;
-  auto callback = [&num_callbacks](Status, std::function<void()>, std::function<void()>) {
-    num_callbacks++;
-  };
-  task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  task_manager_.QueueAndScheduleTask(task2, false, &reply, callback);
-  task_manager_.QueueAndScheduleTask(task3, false, &reply, callback);
-  pool_.TriggerCallbacks();
-
-  // We shouldn't cap anything for zero cpu tasks (and shouldn't crash before
-  // this point).
-  ASSERT_EQ(num_callbacks, 3);
-
-  for (auto &worker : workers) {
-    RayTask buf;
-    task_manager_.TaskFinished(worker, &buf);
-  }
-
-  AssertNoLeaks();
-}
-
-TEST_F(ClusterTaskManagerTestWithoutCPUsAtHead, ZeroCPUNode) {
-  RayTask task = CreateTask({}, /*num_args=*/0, /*args=*/{});
-  RayTask task2 = CreateTask({}, /*num_args=*/0, /*args=*/{});
-  RayTask task3 = CreateTask({}, /*num_args=*/0, /*args=*/{});
-  auto runtime_env_hash = task.GetTaskSpecification().GetRuntimeEnvHash();
-  std::vector<std::shared_ptr<MockWorker>> workers;
-  for (int i = 0; i < 3; i++) {
-    std::shared_ptr<MockWorker> worker =
-        std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-    pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker));
-    pool_.TriggerCallbacks();
-    workers.push_back(worker);
-  }
-  rpc::RequestWorkerLeaseReply reply;
-  int num_callbacks = 0;
-  auto callback = [&num_callbacks](Status, std::function<void()>, std::function<void()>) {
-    num_callbacks++;
-  };
-  task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  task_manager_.QueueAndScheduleTask(task2, false, &reply, callback);
-  task_manager_.QueueAndScheduleTask(task3, false, &reply, callback);
-  pool_.TriggerCallbacks();
-
-  // We shouldn't cap anything for zero cpu tasks (and shouldn't crash before
-  // this point).
-  ASSERT_EQ(num_callbacks, 3);
-
-  for (auto &worker : workers) {
-    RayTask buf;
-    task_manager_.TaskFinished(worker, &buf);
-  }
-  AssertNoLeaks();
-}
-
-/// Test that we exponentially increase the amount of time it takes to increase
-/// the dispatch cap for a scheduling class.
-TEST_F(ClusterTaskManagerTest, SchedulingClassCapIncrease) {
-  auto get_unblocked_worker = [](std::vector<std::shared_ptr<MockWorker>> &workers)
-      -> std::shared_ptr<MockWorker> {
-    for (auto &worker : workers) {
-      if (worker->GetAllocatedInstances() != nullptr && !worker->IsBlocked()) {
-        return worker;
-      }
-    }
-    return nullptr;
-  };
-
-  int64_t UNIT = RayConfig::instance().worker_cap_initial_backoff_delay_ms();
-  std::vector<RayTask> tasks;
-  for (int i = 0; i < 3; i++) {
-    RayTask task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                              /*num_args=*/0, /*args=*/{});
-    tasks.emplace_back(task);
-  }
-
-  rpc::RequestWorkerLeaseReply reply;
-  int num_callbacks = 0;
-  auto callback = [&num_callbacks](Status, std::function<void()>, std::function<void()>) {
-    num_callbacks++;
-  };
-  for (const auto &task : tasks) {
-    task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  }
-
-  auto runtime_env_hash = tasks[0].GetTaskSpecification().GetRuntimeEnvHash();
-  std::vector<std::shared_ptr<MockWorker>> workers;
-  for (int i = 0; i < 3; i++) {
-    std::shared_ptr<MockWorker> worker =
-        std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-    pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker));
-    pool_.TriggerCallbacks();
-    workers.push_back(worker);
-  }
-  task_manager_.ScheduleAndDispatchTasks();
-
-  ASSERT_EQ(num_callbacks, 1);
-
-  current_time_ms_ += UNIT;
-  ASSERT_FALSE(workers.back()->IsBlocked());
-  ASSERT_TRUE(task_manager_.ReleaseCpuResourcesFromUnblockedWorker(
-      get_unblocked_worker(workers)));
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-  task_manager_.ScheduleAndDispatchTasks();
-  ASSERT_EQ(num_callbacks, 2);
-
-  // Since we're increasing exponentially, increasing by a unit show no longer be enough.
-  current_time_ms_ += UNIT;
-  ASSERT_TRUE(task_manager_.ReleaseCpuResourcesFromUnblockedWorker(
-      get_unblocked_worker(workers)));
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-  task_manager_.ScheduleAndDispatchTasks();
-  ASSERT_EQ(num_callbacks, 2);
-
-  // Now it should run
-  current_time_ms_ += UNIT;
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-  task_manager_.ScheduleAndDispatchTasks();
-  ASSERT_EQ(num_callbacks, 3);
-
-  // Let just one task finish.
-  for (auto it = workers.begin(); it != workers.end(); it++) {
-    if (!(*it)->IsBlocked()) {
-      RayTask buf;
-      task_manager_.TaskFinished(*it, &buf);
-      workers.erase(it);
-      break;
-    }
-  }
-
-  current_time_ms_ += UNIT;
-
-  // Now schedule another task of the same scheduling class.
-  RayTask task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                            /*num_args=*/0, /*args=*/{});
-  task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-
-  std::shared_ptr<MockWorker> new_worker =
-      std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-  pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(new_worker));
-  pool_.TriggerCallbacks();
-  workers.push_back(new_worker);
-
-  // It can't run for another 2 units (doesn't increase to 4, because one of
-  // the tasks finished).
-  ASSERT_EQ(num_callbacks, 3);
-
-  current_time_ms_ += 2 * UNIT;
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-  ASSERT_EQ(num_callbacks, 4);
-
-  for (auto &worker : workers) {
-    RayTask buf;
-    task_manager_.TaskFinished(worker, &buf);
-  }
-
-  AssertNoLeaks();
-}
-
-/// Ensure we reset the cap after we've finished executing through the queue.
-TEST_F(ClusterTaskManagerTest, SchedulingClassCapResetTest) {
-  int64_t UNIT = RayConfig::instance().worker_cap_initial_backoff_delay_ms();
-  std::vector<RayTask> tasks;
-  for (int i = 0; i < 2; i++) {
-    RayTask task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                              /*num_args=*/0, /*args=*/{});
-    tasks.emplace_back(task);
-  }
-
-  rpc::RequestWorkerLeaseReply reply;
-  int num_callbacks = 0;
-  auto callback = [&num_callbacks](Status, std::function<void()>, std::function<void()>) {
-    num_callbacks++;
-  };
-  for (const auto &task : tasks) {
-    task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  }
-
-  auto runtime_env_hash = tasks[0].GetTaskSpecification().GetRuntimeEnvHash();
-
-  std::shared_ptr<MockWorker> worker1 =
-      std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-  pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker1));
-  pool_.TriggerCallbacks();
-  task_manager_.ScheduleAndDispatchTasks();
-
-  ASSERT_TRUE(task_manager_.ReleaseCpuResourcesFromUnblockedWorker(worker1));
-  current_time_ms_ += UNIT;
-
-  std::shared_ptr<MockWorker> worker2 =
-      std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-  pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker2));
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-
-  ASSERT_EQ(num_callbacks, 2);
-
-  RayTask buf;
-  task_manager_.TaskFinished(worker1, &buf);
-  task_manager_.TaskFinished(worker2, &buf);
-
-  AssertNoLeaks();
-
-  for (int i = 0; i < 2; i++) {
-    RayTask task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                              /*num_args=*/0, /*args=*/{});
-    task_manager_.QueueAndScheduleTask(task, false, &reply, callback);
-  }
-
-  std::shared_ptr<MockWorker> worker3 =
-      std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-  pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker3));
-  pool_.TriggerCallbacks();
-  task_manager_.ScheduleAndDispatchTasks();
-  ASSERT_EQ(num_callbacks, 3);
-
-  ASSERT_TRUE(task_manager_.ReleaseCpuResourcesFromUnblockedWorker(worker3));
-  current_time_ms_ += UNIT;
-
-  std::shared_ptr<MockWorker> worker4 =
-      std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-  pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker4));
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-
-  ASSERT_EQ(num_callbacks, 4);
-
-  {
-    // Ensure a class of a differenct scheduling class can still be scheduled.
-    RayTask task5 = CreateTask({},
-                               /*num_args=*/0, /*args=*/{});
-    task_manager_.QueueAndScheduleTask(task5, false, &reply, callback);
-    std::shared_ptr<MockWorker> worker5 =
-        std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-    pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker5));
-    task_manager_.ScheduleAndDispatchTasks();
-    pool_.TriggerCallbacks();
-    ASSERT_EQ(num_callbacks, 5);
-    task_manager_.TaskFinished(worker5, &buf);
-  }
-
-  task_manager_.TaskFinished(worker3, &buf);
-  task_manager_.TaskFinished(worker4, &buf);
-
-  AssertNoLeaks();
-}
-
-/// Test that scheduling classes which have reached their running cap start
-/// their timer after the new task is submitted, not before.
-TEST_F(ClusterTaskManagerTest, DispatchTimerAfterRequestTest) {
-  int64_t UNIT = RayConfig::instance().worker_cap_initial_backoff_delay_ms();
-  RayTask first_task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                                  /*num_args=*/0, /*args=*/{});
-
-  rpc::RequestWorkerLeaseReply reply;
-  int num_callbacks = 0;
-  auto callback = [&num_callbacks](Status, std::function<void()>, std::function<void()>) {
-    num_callbacks++;
-  };
-  task_manager_.QueueAndScheduleTask(first_task, false, &reply, callback);
-
-  auto runtime_env_hash = first_task.GetTaskSpecification().GetRuntimeEnvHash();
-  std::vector<std::shared_ptr<MockWorker>> workers;
-  for (int i = 0; i < 3; i++) {
-    std::shared_ptr<MockWorker> worker =
-        std::make_shared<MockWorker>(WorkerID::FromRandom(), 1234, runtime_env_hash);
-    pool_.PushWorker(std::static_pointer_cast<WorkerInterface>(worker));
-    pool_.TriggerCallbacks();
-    workers.push_back(worker);
-  }
-  task_manager_.ScheduleAndDispatchTasks();
-
-  ASSERT_EQ(num_callbacks, 1);
-
-  RayTask second_task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                                   /*num_args=*/0, /*args=*/{});
-  task_manager_.QueueAndScheduleTask(second_task, false, &reply, callback);
-  pool_.TriggerCallbacks();
-
-  /// Can't schedule yet due to the cap.
-  ASSERT_EQ(num_callbacks, 1);
-  for (auto &worker : workers) {
-    if (worker->GetAllocatedInstances() && !worker->IsBlocked()) {
-      task_manager_.ReleaseCpuResourcesFromUnblockedWorker(worker);
-    }
-  }
-
-  current_time_ms_ += UNIT;
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-
-  ASSERT_EQ(num_callbacks, 2);
-  for (auto &worker : workers) {
-    if (worker->GetAllocatedInstances() && !worker->IsBlocked()) {
-      task_manager_.ReleaseCpuResourcesFromUnblockedWorker(worker);
-    }
-  }
-
-  /// A lot of time passes, definitely more than the timeout.
-  current_time_ms_ += 100000 * UNIT;
-
-  RayTask third_task = CreateTask({{ray::kCPU_ResourceLabel, 8}},
-                                  /*num_args=*/0, /*args=*/{});
-  task_manager_.QueueAndScheduleTask(third_task, false, &reply, callback);
-  pool_.TriggerCallbacks();
-
-  /// We still can't schedule the third task since the timer doesn't start
-  /// until after the task is queued.
-  ASSERT_EQ(num_callbacks, 2);
-
-  current_time_ms_ += 2 * UNIT;
-  task_manager_.ScheduleAndDispatchTasks();
-  pool_.TriggerCallbacks();
-
-  ASSERT_EQ(num_callbacks, 3);
-
-  for (auto &worker : workers) {
-    RayTask buf;
-    task_manager_.TaskFinished(worker, &buf);
-  }
-
-  AssertNoLeaks();
-}
-
 // Regression test for https://github.com/ray-project/ray/issues/16935:
 // When a task requires 1 CPU and is infeasible because head node has 0 CPU,
 // make sure the task's resource demand is reported.
@@ -2112,17 +1660,15 @@ TEST_F(ClusterTaskManagerTestWithoutCPUsAtHead, OneCpuInfeasibleTask) {
     const auto &resource_load_by_shape = data.resource_load_by_shape();
     ASSERT_EQ(resource_load_by_shape.resource_demands().size(), demand_types[i]);
 
-    // Assert that the one-cpu fields are correct.
-    bool one_cpu_found = false;
+    // Find the 1-cpu task and make sure the demand quantities are correct.
+    bool found = false;
     for (const auto &demand : resource_load_by_shape.resource_demands()) {
-      if (demand.shape().at("CPU") == 1) {
-        ASSERT_FALSE(one_cpu_found);
-        one_cpu_found = true;
+      if (demand.shape().size() == 1 && demand.shape().at("CPU") == 1) {
+        found = true;
         EXPECT_EQ(demand.num_infeasible_requests_queued(), num_infeasible_1cpu[i]);
-        ASSERT_EQ(demand.shape().size(), 1);
       }
     }
-    ASSERT_TRUE(one_cpu_found);
+    ASSERT_TRUE(found);
   }
 }
